+++ resolwe_bio/docker_images/dnaseq/packages-manual/gatk.sh
@@ -5,8 +5,8 @@
 download_and_verify \
     broadinstitute \
     gatk \
-    4.2.0.0 \
-    dd11cc8e3bc7a23c2c226366428f0908c902765eabbc1c641c736c06b80aaf78 \
+    4.2.4.1 \
+    2dc007f534cca1e29253864601e480ce24ce376b4b2bbd7a491571f9ecedc2e2 \
     https://github.com/broadinstitute/gatk/releases/download/\${version}/gatk-\${version}.zip \
     gatk-\${version}