+++ resolwe_bio/docker_images/dnaseq/packages-manual/r-packages.sh
@@ -2,16 +2,14 @@
 
 . /var/cache/build/packages-manual/common.sh
 
-# These steps need to be performed because certain packages produce a "database corrupt error". The solution is to
-# restart R, but this is not feasible during the docker built or building of dependencies. This chunk also installs
-# packages from bioconductor. This entire step should be considered temporary until we can get packrat to work for
-# packages from all universes (CRAN, Bioconductor).
-Rscript --slave --no-save --no-restore-history -e "install.packages('assertthat')"
+# Install R packages.
+# Install renv which is a pre-requisite to installing all other packages in next chunk.
+Rscript --slave --no-save --no-restore-history -e "install.packages('renv', repos = 'https://cloud.r-project.org')"
+Rscript --slave --no-save --no-restore-history -e "renv::settings\$bioconductor.version('3.13')"
 
-Rscript --slave --no-save --no-restore-history -e \
-"source('https://bioconductor.org/biocLite.R'); \
-BiocInstaller::biocLite(c('IRanges', 'GenomicRanges', 'Rsamtools'))\
-"
+sed -e 's/^#.*$//g' -e '/^$/d' /var/cache/build/packages-r.txt | \
+    Rscript --slave --no-save --no-restore-history \
+        -e "renv::install(readLines('stdin'), repos='https://cloud.r-project.org')"
 
 # XXX: This is unverifiable and thus may compromise the whole image.
 # XXX: Use notary (https://github.com/ropenscilabs/notary) when ready.